refine action_data_to_json related interface

+++ libraries/eosiolib/capi/eosio/action.h
@@ -201,7 +201,7 @@ void set_action_return_value(void *return_value, size_t size);
  * @return the initialized convertor id. Used by final_action_data_to_json()
  */
 __attribute__((eosio_wasm_import))
-uint64_t init_action_data_to_json(uint64_t contract_name, uint64_t action_name, const char* action_data, uint32_t action_data_size, uint64_t* json_size);
+uint64_t init_action_data_to_json(uint64_t contract_name, uint64_t action_name, const char* action_data, uint32_t action_data_size, uint32_t* json_size);
 
 /**
  * Finalize the conversation from action data to JSON compact format.
@@ -213,7 +213,7 @@ uint64_t init_action_data_to_json(uint64_t contract_name, uint64_t action_name, 
  *
  */
 __attribute__((eosio_wasm_import))
-void final_action_data_to_json(const uint64_t convertor_id, char* json_output, uint64_t json_size);
+void final_action_data_to_json(const uint64_t convertor_id, char* json_output, uint32_t json_size);
 
 
 #ifdef __cplusplus
